Checking for collinearity between JCR and AIS.
To determine if it was appropriate to include both AIS and JCR in our final model, I updated the variables we check for collinearity in this script, dropping APC in the process. I also added a command to load the package `tidyverse` because this script uses a command from that package.